rename: mkdir config, order, voucher and move files
