+++ src/main/java/org/prgrms/kdtspringw1d1/order/Order.java
@@ -1,4 +1,6 @@
-package org.prgrms.kdtspringw1d1;
+package org.prgrms.kdtspringw1d1.order;
+
+import org.prgrms.kdtspringw1d1.voucher.Voucher;
 
 import java.util.List;
 import java.util.UUID;
