+++ src/main/java/org/prgrms/kdtspringw1d1/order/OrderContext.java
@@ -1,4 +1,8 @@
-package org.prgrms.kdtspringw1d1;
+package org.prgrms.kdtspringw1d1.order;
+
+import org.prgrms.kdtspringw1d1.voucher.Voucher;
+import org.prgrms.kdtspringw1d1.voucher.VoucherRepository;
+import org.prgrms.kdtspringw1d1.voucher.VoucherService;
 
 import java.util.Optional;
 import java.util.UUID;
