+++ src/main/java/org/prgrms/kdtspringw1d1/order/OrderService.java
@@ -1,6 +1,7 @@
-package org.prgrms.kdtspringw1d1;
+package org.prgrms.kdtspringw1d1.order;
 
-import java.util.Collections;
+import org.prgrms.kdtspringw1d1.voucher.VoucherService;
+
 import java.util.List;
 import java.util.UUID;
